requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
drf-spectacular==0.26.5

# Database
psycopg

# Security & Authentication
argon2-cffi==23.1.0

# Admin Theme
django-admin-interface==0.27.0
django-flat-theme==1.1.4
django-colorfield==0.14.0


# Environment & Configuration
python-decouple==3.8

# Development & Testing
pytest==7.4.3
pytest-django==4.6.0
coverage==7.3.2
black==23.10.1
flake8==6.1.0
isort==5.12.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Documentation
drf-spectacular-sidecar==2023.10.1
